move icon to seperate folder in example
