+++ example/icon/iconunix.go
@@ -2,9 +2,9 @@
 
 // File generated by 2goarray (http://github.com/cratonica/2goarray)
 
-package main
+package icon
 
-var iconData []byte = []byte {
+var Data []byte = []byte {
 	0x89, 0x50, 0x4e, 0x47, 0x0d, 0x0a, 0x1a, 0x0a, 0x00, 0x00, 0x00, 0x0d, 
 	0x49, 0x48, 0x44, 0x52, 0x00, 0x00, 0x00, 0x20, 0x00, 0x00, 0x00, 0x20, 
 	0x08, 0x06, 0x00, 0x00, 0x00, 0x73, 0x7a, 0x7a, 0xf4, 0x00, 0x00, 0x00,
