+++ example/icon/make_icon.sh
@@ -28,7 +28,7 @@ OUTPUT=iconunix.go
 echo Generating $OUTPUT
 echo "//+build linux darwin" > $OUTPUT
 echo >> $OUTPUT
-cat "$1" | $GOPATH/bin/2goarray iconData main >> $OUTPUT
+cat "$1" | $GOPATH/bin/2goarray Data icon >> $OUTPUT
 if [ $? -ne 0 ]; then
     echo Failure generating $OUTPUT
     exit
